chore(typings): move NormalizedCache types to apollo-cache-inmemory
Addresses comment: https://github.com/apollographql/apollo-client/pull/2362#discussion_r147994949
