+++ packages/apollo-cache-inmemory/src/types.ts
@@ -1,13 +1,44 @@
 import { DocumentNode } from 'graphql';
 import { FragmentMatcher } from 'graphql-anywhere';
-import {
-  Transaction,
-  NormalizedCacheObject,
-  NormalizedCache,
-} from 'apollo-cache';
-import { IdValue } from 'apollo-utilities';
+import { Transaction } from 'apollo-cache';
+import { IdValue, StoreValue } from 'apollo-utilities';
+import { NormalizedCacheObject, NormalizedCache } from './types';
 
 export type IdGetter = (value: Object) => string | null | undefined;
+
+/**
+ * This is an interface used to access, set and remove
+ * StoreObjects from the cache
+ */
+export interface NormalizedCache {
+  get(dataId: string): StoreObject;
+  set(dataId: string, value: StoreObject): void;
+  delete(dataId: string): void;
+  clear(): void;
+
+  // non-Map elements:
+  /**
+   * returns an Object with key-value pairs matching the contents of the store
+   */
+  toObject(): NormalizedCacheObject;
+  /**
+   * replace the state of the store
+   */
+  replace(newData: NormalizedCacheObject): void;
+}
+
+/**
+ * This is a normalized representation of the Apollo query result cache. It consists of
+ * a flattened representation of query result trees.
+ */
+export interface NormalizedCacheObject {
+  [dataId: string]: StoreObject;
+}
+
+export interface StoreObject {
+  __typename?: string;
+  [storeFieldKey: string]: StoreValue;
+}
 
 export type NormalizedCacheFactory = (
   seed?: NormalizedCacheObject,
@@ -105,8 +136,3 @@ export type CacheResolverMap = {
 // backwards compat
 export type CustomResolver = CacheResolver;
 export type CustomResolverMap = CacheResolverMap;
-export {
-  NormalizedCacheObject,
-  NormalizedCache,
-  StoreObject,
-} from 'apollo-cache';
