+++ packages/apollo-cache/src/types/index.ts
@@ -1,3 +1,2 @@
 export * from './DataProxy';
 export * from './Cache';
-export * from './NormalizedCache';
